added date and move sorting interface to another file
